build(Simulator): move to simulator/
