+++ CMakeLists.txt
@@ -49,9 +49,10 @@ set(SOURCE_FILES
 	"src/rocket/Engine.cpp"
 	"src/rocket/RocketSpec.cpp"
 
-	"src/solver/DetailSimulator.cpp"
-	"src/solver/ScatterSimulator.cpp"
-	"src/solver/Simulator.cpp"
+	"src/simulator/DetailSimulator.cpp"
+	"src/simulator/ScatterSimulator.cpp"
+	"src/simulator/Simulator.cpp"
+
 	"src/solver/Solver.cpp"
 
 	"src/utils/JsonUtils.cpp"
